move to shared encoding module
